+++ .gitlab/conf/host-configs/blueos_3_ppc64le_ib/nvcc_clang_coral_2017_08_31.cmake
@@ -1,9 +1,0 @@
-##############################################################################
-# Copyright (c) 2016-19, Lawrence Livermore National Security, LLC and Umpire
-# project contributors. See the COPYRIGHT file for details.
-#
-# SPDX-License-Identifier: (MIT)
-##############################################################################
-set(CMAKE_CXX_COMPILER "/usr/tce/packages/clang/clang-coral-2017.08.31/bin/clang++" CACHE PATH "")
-set(CMAKE_C_COMPILER "/usr/tce/packages/clang/clang-coral-2017.08.31/bin/clang" CACHE PATH "")
-set(CMAKE_Fortran_COMPILER "/usr/tce/packages/xl/xl-beta-2017.09.13/bin/xlf2003_r" CACHE PATH "")