+++ src/umpire/op/MemoryOperationRegistry.cpp
@@ -1,5 +1,5 @@
 //////////////////////////////////////////////////////////////////////////////
-// Copyright (c) 2016-19, Lawrence Livermore National Security, LLC and Umpire
+// Copyright (c) 2016-20, Lawrence Livermore National Security, LLC and Umpire
 // project contributors. See the COPYRIGHT file for details.
 //
 // SPDX-License-Identifier: (MIT)
@@ -32,11 +32,6 @@
 #include "umpire/op/CudaAdviseUnsetAccessedByOperation.hpp"
 #include "umpire/op/CudaAdviseUnsetPreferredLocationOperation.hpp"
 #include "umpire/op/CudaAdviseUnsetReadMostlyOperation.hpp"
-#endif
-
-#if defined(UMPIRE_ENABLE_HCC)
-#include "umpire/op/RocmCopyOperation.hpp"
-#include "umpire/op/RocmMemsetOperation.hpp"
 #endif
 
 #if defined(UMPIRE_ENABLE_HIP)
@@ -92,7 +87,6 @@ MemoryOperationRegistry::MemoryOperationRegistry() noexcept
       "MOVE",
       std::make_pair(Platform::cuda, Platform::cpu),
       std::make_shared<NumaMoveOperation>());
-// NOTE: Add cpu<->rocm pairs here when needed
 #endif
 
 #if defined(UMPIRE_ENABLE_CUDA)
@@ -165,33 +159,6 @@ MemoryOperationRegistry::MemoryOperationRegistry() noexcept
       "PREFETCH",
       std::make_pair(Platform::cuda, Platform::cuda),
       std::make_shared<CudaMemPrefetchOperation>());
-#endif
-
-#if defined(UMPIRE_ENABLE_HCC)
-  registerOperation(
-      "COPY",
-      std::make_pair(Platform::rocm, Platform::cpu),
-      std::make_shared<RocmCopyOperation>());
-
-  registerOperation(
-      "COPY",
-      std::make_pair(Platform::cpu, Platform::rocm),
-      std::make_shared<RocmCopyOperation>());
-
-  registerOperation(
-      "COPY",
-      std::make_pair(Platform::rocm, Platform::rocm),
-      std::make_shared<RocmCopyOperation>());
-
-  registerOperation(
-      "MEMSET",
-      std::make_pair(Platform::rocm, Platform::rocm),
-      std::make_shared<RocmMemsetOperation>());
-
-  registerOperation(
-      "REALLOCATE",
-      std::make_pair(Platform::rocm, Platform::rocm),
-      std::make_shared<GenericReallocateOperation>());
 #endif
 
 #if defined(UMPIRE_ENABLE_HIP)