+++ tests/integration/operation_tests.cpp
@@ -1,5 +1,5 @@
 //////////////////////////////////////////////////////////////////////////////
-// Copyright (c) 2016-19, Lawrence Livermore National Security, LLC and Umpire
+// Copyright (c) 2016-20, Lawrence Livermore National Security, LLC and Umpire
 // project contributors. See the COPYRIGHT file for details.
 //
 // SPDX-License-Identifier: (MIT)
@@ -282,6 +282,83 @@ class ReallocateTest :
 {
 };
 
+// This value is such that the 64Kb limit on device constant memory is not hit
+// in check_alloc_realloc_free when reallocating to 3 * SIZE.
+constexpr int SIZE = 5345;
+
+TEST_P(ReallocateTest, ReallocateSweep)
+{
+  umpire::ResourceManager& rm = umpire::ResourceManager::getInstance();
+  const bool hostAccessible = source_allocator->getId() == rm.getAllocator("HOST").getId();
+
+  for ( int size = 0 ; size <= SIZE ; size = size * 2 + 1 ) {
+    int buffer_size = size;
+    int* buffer = nullptr;
+
+    ASSERT_NO_THROW({
+      buffer = static_cast<int*>(
+          source_allocator->allocate(buffer_size*sizeof(*buffer)));
+    });
+
+    ASSERT_EQ(source_allocator->getId(), rm.getAllocator(buffer).getId());
+
+    if ( hostAccessible ) {
+      // Populate the buffer.
+      for ( int i = 0 ; i < buffer_size ; ++i ) {
+        buffer[ i ] = i;
+      }
+
+      // Check the values.
+      for ( int i = 0 ; i < buffer_size ; ++i ) {
+        ASSERT_EQ( buffer[ i ], i );
+      }
+    }
+
+    // Reallocate to a larger size.
+    buffer_size *= 3;
+    ASSERT_NO_THROW({
+      buffer = static_cast<int*>(rm.reallocate( buffer, buffer_size*sizeof(*buffer) ));
+    });
+    if (buffer_size > 0) {
+      ASSERT_EQ(source_allocator->getId(), rm.getAllocator(buffer).getId());
+    }
+
+    if ( hostAccessible ) {
+      // Populate the new values.
+      for ( int i = size ; i < buffer_size ; ++i ) {
+        buffer[ i ] = i;
+      }
+
+      // Check all the values.
+      for ( int i = 0 ; i < buffer_size ; ++i ) {
+        EXPECT_EQ( buffer[ i ], i );
+      }
+    }
+
+    // Reallocate to a smaller size.
+    buffer_size /= 5;
+    ASSERT_NO_THROW({
+      buffer = static_cast<int*>(rm.reallocate( buffer, buffer_size*sizeof(*buffer) ));
+    });
+    if (buffer_size > 0) {
+      ASSERT_EQ(source_allocator->getId(), rm.getAllocator(buffer).getId());
+    }
+
+    if ( hostAccessible ) {
+      // Check all the values.
+      for ( int i = 0 ; i < buffer_size ; ++i ) {
+        EXPECT_EQ( buffer[ i ], i );
+      }
+    }
+
+    // Free
+    ASSERT_NO_THROW({
+      source_allocator->deallocate( buffer );
+    });
+    //EXPECT_TRUE( buffer == nullptr );
+  }
+}
+
 TEST_P(ReallocateTest, Reallocate)
 {
   umpire::ResourceManager& rm = umpire::ResourceManager::getInstance();
@@ -540,11 +617,10 @@ TEST_P(ReallocateTest, ReallocateWithAllocator)
 
 TEST_P(ReallocateTest, ReallocateWithAllocatorFail)
 {
-  umpire::ResourceManager& rm = umpire::ResourceManager::getInstance();
-
   if (source_allocator->getId() == dest_allocator->getId()) {
     SUCCEED();
   } else {
+    umpire::ResourceManager& rm = umpire::ResourceManager::getInstance();
     ASSERT_THROW(
         rm.reallocate(source_array, m_size, *dest_allocator),
         umpire::util::Exception);